introduce params for email body template and email subject - funded by Mediwe

+++ CRM/Pdfapi/Upgrader.php
@@ -13,7 +13,16 @@ class CRM_Pdfapi_Upgrader extends CRM_Pdfapi_Upgrader_Base {
    * Add CiviRule Action when installing
    */
   public function install() {
-    $this->executeSqlFile('sql/insertSendPDFAction.sql');
+    // if CiviRules installed
+    try {
+      $extensions = civicrm_api3('Extension', 'get');
+      foreach($extensions['values'] as $ext) {
+        if ($ext['key'] == 'org.civicoop.civirules' &&$ext['status'] == 'installed') {
+          $this->executeSqlFile('sql/insertSendPDFAction.sql');
+        }
+      }
+    } catch (CiviCRM_API3_Exception $ex) {
+    }
   }
 
   /**
@@ -21,12 +30,67 @@ class CRM_Pdfapi_Upgrader extends CRM_Pdfapi_Upgrader_Base {
    */
   public function upgrade_1001() {
     $this->ctx->log->info('Applying update 1001 (remove managed entity');
-    if (CRM_Core_DAO::checkTableExists('civicrm_managed')) {
-      $query = 'DELETE FROM civicrm_managed WHERE module = %1 AND entity_type = %2';
-      CRM_Core_DAO::executeQuery($query, array(
-        1 => array('org.civicoop.pdfapi', 'String'),
-        2 => array('CiviRuleAction', 'String'),
-      ));
+    // if CiviRules installed
+    try {
+      $extensions = civicrm_api3('Extension', 'get');
+      foreach($extensions['values'] as $ext) {
+        if ($ext['key'] == 'org.civicoop.civirules' &&$ext['status'] == 'installed') {
+          if (CRM_Core_DAO::checkTableExists('civicrm_managed')) {
+            $query = 'DELETE FROM civicrm_managed WHERE module = %1 AND entity_type = %2';
+            CRM_Core_DAO::executeQuery($query, array(
+              1 => array('org.civicoop.pdfapi', 'String'),
+              2 => array('CiviRuleAction', 'String'),
+            ));
+          }
+        }
+      }
+    } catch (CiviCRM_API3_Exception $ex) {
+    }
+    return TRUE;
+  }
+
+  /**
+   * update action params in existing usage for body_template_id and email_subject
+   */
+  public function upgrade_1002() {
+    $this->ctx->log->info('Applying update 1002 (introduce email body template and email subject');
+    // if CiviRules installed
+    try {
+      $extensions = civicrm_api3('Extension', 'get');
+      foreach($extensions['values'] as $ext) {
+        if ($ext['key'] == 'org.civicoop.civirules' &&$ext['status'] == 'installed') {
+
+          $query = "SELECT id FROM civirule_action WHERE name = %1 AND class_name = %2";
+          $actionId = CRM_Core_DAO::singleValueQuery($query, array(
+            1 => array('pdfapi_send', 'String'),
+            2 => array('CRM_Pdfapi_CivirulesAction', 'String'),
+          ));
+          if ($actionId) {
+            // update params in rule action if any present
+            $query = "SELECT id, action_params FROM civirule_rule_action WHERE action_id = %1";
+            $dao = CRM_Core_DAO::executeQuery($query, array(1 => array((int) $actionId, 'Integer')));
+            while ($dao->fetch()) {
+              $actionParams = unserialize($dao->action_params);
+              $updateRuleAction = FALSE;
+              if (!isset($actionParams['body_template_id'])) {
+                $actionParams['body_template_id'] = "";
+                $updateRuleAction = TRUE;
+              }
+              if (!isset($actionParams['email_subject'])) {
+                $actionParams['email_subject'] = "";
+                $updateRuleAction = TRUE;
+              }
+              if ($updateRuleAction) {
+                $ruleAction = new CRM_Civirules_BAO_RuleAction();
+                $ruleAction->id = $dao->id;
+                $ruleAction->action_params = serialize($actionParams);
+                $ruleAction->save();
+              }
+            }
+          }
+        }
+      }
+    } catch (CiviCRM_API3_Exception $ex) {
     }
     return TRUE;
   }
